lambda/pyproject.toml
[tool.poetry]
name = "lambda"
version = "0.1.0"
description = ""
authors = ["Rinku <shark4rkcoding@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
boto3 = "^1.37.37"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
